refactor cadvisor manager
prep to be ready for other container engines
